requirements.txt
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit>=1.28.0
pymongo>=4.6.0
pathlib
markdown>=3.5.0
pdfkit>=1.0.0
pandas>=2.0.0
reportlab>=4.0.0
beautifulsoup4>=4.12.0
